Edit waitinglist class model,repo,service, controller.

+++ src/main/java/com/aiu/library/datastructures/BillingList.java
@@ -1,7 +1,5 @@
 package com.aiu.library.datastructures;
 
-
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -10,34 +8,28 @@ import com.aiu.library.model.Billing;
 
 public class BillingList {
 
-    
     private List<Billing> billings;
 
-    
     public BillingList() {
         this.billings = new ArrayList<>();
     }
 
-    
     public void addBilling(Billing billing) {
         if (billing != null) {
             this.billings.add(billing);
         }
     }
 
-    
-    public boolean removeBilling(Long id) {
+    public boolean removeBilling(int id) {
         return this.billings.removeIf(billing -> billing.getId().equals(id));
     }
 
-    
     public Optional<Billing> getBillingById(Long id) {
         return this.billings.stream()
                 .filter(billing -> billing.getId().equals(id))
                 .findFirst();
     }
 
-    
     public Optional<Billing> getBillingByMemberId(Long memberId) {
         return this.billings.stream()
                 .filter(billing -> billing.getMemberID().equals(memberId))
@@ -45,7 +37,7 @@ public class BillingList {
     }
 
     public List<Billing> getAllBillings() {
-        return new ArrayList<>(this.billings); 
+        return new ArrayList<>(this.billings);
     }
 
     public Double getTotalFines() {
@@ -54,20 +46,16 @@ public class BillingList {
                 .sum();
     }
 
-    
     public int size() {
         return this.billings.size();
     }
 
-    
     public boolean isEmpty() {
         return this.billings.isEmpty();
     }
 
-    
     public void clear() {
         this.billings.clear();
     }
 
 }
-
